+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT UP
 
 - Monster feature effects: improve parsing of invisibility to handle situations for things like the Poltergeist which are always invisible, and the Pixie which has Superior Invisibility.
+- Active Aura Effect fixes: Changes in DAE caused Active Aura effects to stop working correctly.
 
 # 3.5.0